refactor(shared): move constants from type files to constants directory
Separate runtime constants from type definitions following single-responsibility
principle. Type files now contain only types/interfaces, while constants live in
shared/src/constants/. Also removes dead legacy template aliases and unexports
internal-only template constants.
