+++ packages/manifest/frontend/src/components/theme-editor/VariableControlGroup.tsx
@@ -1,7 +1,7 @@
 import type { ThemeVariables } from '@manifest/shared';
 import { ColorPickerControl } from './ColorPickerControl';
 import { RadiusControl } from './RadiusControl';
-import type { ThemeVariableGroup } from './types';
+import type { ThemeVariableGroup } from '@manifest/shared';
 
 interface VariableControlGroupProps {
   /** Group configuration */
