+++ packages/manifest/shared/src/constants/platform.ts
@@ -1,0 +1,9 @@
+import type { PreviewPreferences } from '../types/platform.js';
+
+/**
+ * Default preview preferences
+ */
+export const DEFAULT_PREVIEW_PREFERENCES: PreviewPreferences = {
+  platformStyle: 'chatgpt',
+  themeMode: 'light',
+};
