+++ packages/manifest/shared/src/types/analytics.ts
@@ -8,14 +8,6 @@ export type AnalyticsTimeRange = '24h' | '7d' | '30d' | '3mo';
 /** Metric types for chart display selection */
 export type SelectedMetric = 'executions' | 'uniqueUsers' | 'completionRate' | 'avgDuration';
 
-/** Time range display labels */
-export const TIME_RANGE_LABELS: Record<AnalyticsTimeRange, string> = {
-  '24h': 'Last 24 Hours',
-  '7d': 'Last 7 Days',
-  '30d': 'Last 30 Days',
-  '3mo': 'Last 3 Months',
-};
-
 /** Single data point for the area chart */
 export interface ChartDataPoint {
   /** Time bucket (ISO string or formatted date) */
