+++ packages/manifest/shared/src/types/platform.ts
@@ -16,10 +16,3 @@ export interface PreviewPreferences {
   themeMode: ThemeMode;
 }
 
-/**
- * Default preview preferences
- */
-export const DEFAULT_PREVIEW_PREFERENCES: PreviewPreferences = {
-  platformStyle: 'chatgpt',
-  themeMode: 'light',
-};
